fix: update return code
Currently, ua fix will always return 0 no matter what happens
during the execution of the command. We are updating the command
to only return 0 if the system ends up in a non-vulnerable state.
Meaning that either the CVE/USN does not affect the system, the
packages required by the fix are already installed or that
we have successfully installed all of the packages necessary for the
fix.