postinst: hardcode python binary and PYTHONPATH
When the user's environment is set up to use a nonstandard python
installation, our postinst would fail. By specifying the exact python
binary to use, we avoid this problem. By also setting PYTHONPATH to the
location where our uaclient library is installed, we increase the
chances that we will succeed even if /usr/bin/python3 is overwritten.

If /usr/bin/python3 is overwritten with a different version of python,
whose standard library is incompatible with the expected version of
python, it is possible postinst could still fail.

Separately, if /usr/bin/python3 is overwritten with a different version
of python, the ua executable _will_ still fail.

LP: #1930121